+++ CHANGELOG.md
@@ -7,9 +7,16 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ## Unreleased
 
+### Added
+
+- application-wide and query-specific ways to define default prefixes 
+
 ### Fixed
 
 - bug in the lexer grammar which caused a scanner error on lowercase `distinct`
+
+[Compare v0.2.1...HEAD](https://github.com/marcelotto/sparql-ex/compare/v0.2.1...HEAD)
+
 
 
 ## 0.2.1 - 2018-09-17